requirements.txt
agora_token_builder==1.0.0
aioredis==1.3.1
amqp==5.1.1
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.1.1
Automat==20.2.0
Babel==2.9.1
backports.zoneinfo==0.2.1;python_version<"3.9"
billiard==4.2.1
CacheControl==0.12.11
cachetools==5.2.0
celery==5.4.0
certifi==2021.10.8
cffi==1.17.1
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.11
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
daphne==3.0.2
Django==4.0.2
django-admin-autocomplete-filter==0.7.1
django-cors-headers==3.10.1
django-environ==0.8.1
django-filter==21.1
django-modeltranslation==0.17.5
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
firebase-admin==6.0.1
generics==3.6.0
gevent==24.11.1
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.2
google-cloud-storage==2.6.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
greenlet==3.1.1
grpcio==1.69.0
grpcio-status==1.50.0
gunicorn==20.1.0
hiredis==3.1.0
httplib2==0.21.0
hyperlink==21.0.0
i18n==0.2
idna==3.3
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
kombu==5.4.2
MarkupSafe==2.1.0
msgpack==1.0.3
openapi-codec==1.3.2
packaging==21.3
phonenumbers==8.12.43
Pillow==8.4.0;python_version<"3.10"
Pillow==11.1.0;python_version>"3.10"
prompt-toolkit==3.0.36
proto-plus==1.22.1
protobuf==4.21.9
psycopg2==2.9.10
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
python-dateutil==2.9.0.post0
pytz==2021.3
redis==4.4.0
requests==2.27.1
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
service-identity==21.1.0
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
Twisted==22.1.0
txaio==21.2.1
typing_extensions==4.1.1
tzdata==2024.2
uritemplate==4.1.1
urllib3==1.26.8
vine==5.1.0
wcwidth==0.2.5
websockets==12.0
zope.event==4.6
zope.interface==5.4.0
